refactor mapping function handling
